requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
requests~=2.28.1
